added tests for login and signup
